+++ src/pages/login/components/LoginFormSection.tsx
@@ -1,47 +1,23 @@
 import styled from "@emotion/styled";
-import { useEffect, useState } from "react";
-import LoginInput from "./LoginInput";
+import LoginInput from "@/pages/login/components/LoginInput";
+import { useLoginForm } from "@/hooks/useLoginForm";
 
 type Props = {
   onLogin: () => void;
 };
 
-const validateEmail = (email: string) => {
-  if (!email) return "아이디를 입력해주세요.";
-  const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-  if (!emailRegex.test(email)) return "아이디는 이메일 형식으로 입력해주세요.";
-  return "";
-};
-
-const validatePassword = (password: string) => {
-  if (!password) return "비밀번호를 입력해주세요.";
-  if (password.length < 8) return "비밀번호는 최소 8글자 이상이어야 합니다.";
-  return "";
-};
-
 export default function LoginFormSection({ onLogin }: Props) {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-
-  const [emailError, setEmailError] = useState("");
-  const [passwordError, setPasswordError] = useState("");
-
-  const [isButtonEnabled, setIsButtonEnabled] = useState(false);
-
-  const handleEmailBlur = () => {
-    setEmailError(validateEmail(email));
-  };
-
-  const handlePasswordBlur = () => {
-    setPasswordError(validatePassword(password));
-  };
-
-  useEffect(() => {
-    const emailValid = validateEmail(email) === "";
-    const passwordValid = validatePassword(password) === "";
-
-    setIsButtonEnabled(emailValid && passwordValid);
-  }, [email, password]);
+  const {
+    email,
+    password,
+    emailError,
+    passwordError,
+    isButtonEnabled,
+    setEmail,
+    setPassword,
+    handleEmailBlur,
+    handlePasswordBlur,
+  } = useLoginForm();
 
   return (
     <Form onSubmit={(e) => e.preventDefault()}>
